Convert forceVisualforceComponentCreate
